Split render_wasm.rs and render_wasm_p2.rs to clear max-lines (keep mod registry/tests declarations in render_wasm.rs itself for correct path resolution)
